+++ order-microservice/src/main/java/com/metao/book/order/application/cart/UpdateCartItemQtyDTO.java
@@ -1,4 +1,6 @@
 package com.metao.book.order.application.cart;
 
-public record UpdateCartItemQtyDTO(int quantity) {
+import java.math.BigDecimal;
+
+public record UpdateCartItemQtyDTO(BigDecimal quantity) {
 }
